Move signature package to eks-anywhere
