+++ pkg/signature/manifest.go
@@ -25,14 +25,30 @@ import (
 	"strings"
 	"text/template"
 
+	"github.com/aws/aws-sdk-go-v2/config"
 	"github.com/aws/aws-sdk-go-v2/service/kms"
 	"github.com/aws/aws-sdk-go-v2/service/kms/types"
 	"github.com/itchyny/gojq"
 	"sigs.k8s.io/yaml"
 
 	anywherev1alpha1 "github.com/aws/eks-anywhere/release/api/v1alpha1"
-	"github.com/aws/eks-anywhere/release/cli/pkg/clients"
-	"github.com/aws/eks-anywhere/release/cli/pkg/constants"
+)
+
+const (
+	// DefaultRegion used to create KMS client.
+	DefaultRegion = "us-west-2"
+	// KmsKey alias.
+	KmsKey = "arn:aws:kms:us-west-2:857151390494:alias/signingEKSABundlesKey"
+
+	// SignatureAnnotation applied to the bundle during bundle manifest signing.
+	SignatureAnnotation = "anywhere.eks.amazonaws.com/signature"
+	// ExcludesAnnotation applied to the bundle during bundle manifest signing.
+	ExcludesAnnotation = "anywhere.eks.amazonaws.com/excludes"
+
+	// Excludes is a base64-encoded, newline-delimited list of JSON/YAML paths to remove
+	// from the Bundles manifest prior to computing the digest. You can add or remove
+	// fields depending on your signing requirements.
+	Excludes = "LnNwZWMudmVyc2lvbnNCdW5kbGVzW10uYm9vdHN0cmFwCi5zcGVjLnZlcnNpb25zQnVuZGxlc1tdLmJvdHRsZXJvY2tldEhvc3RDb250YWluZXJzCi5zcGVjLnZlcnNpb25zQnVuZGxlc1tdLmNlcnRNYW5hZ2VyCi5zcGVjLnZlcnNpb25zQnVuZGxlc1tdLmNpbGl1bQouc3BlYy52ZXJzaW9uc0J1bmRsZXNbXS5jbG91ZFN0YWNrCi5zcGVjLnZlcnNpb25zQnVuZGxlc1tdLmNsdXN0ZXJBUEkKLnNwZWMudmVyc2lvbnNCdW5kbGVzW10uY29udHJvbFBsYW5lCi5zcGVjLnZlcnNpb25zQnVuZGxlc1tdLmRvY2tlcgouc3BlYy52ZXJzaW9uc0J1bmRsZXNbXS5la3NhCi5zcGVjLnZlcnNpb25zQnVuZGxlc1tdLmV0Y2RhZG1Cb290c3RyYXAKLnNwZWMudmVyc2lvbnNCdW5kbGVzW10uZXRjZGFkbUNvbnRyb2xsZXIKLnNwZWMudmVyc2lvbnNCdW5kbGVzW10uZmx1eAouc3BlYy52ZXJzaW9uc0J1bmRsZXNbXS5oYXByb3h5Ci5zcGVjLnZlcnNpb25zQnVuZGxlc1tdLmtpbmRuZXRkCi5zcGVjLnZlcnNpb25zQnVuZGxlc1tdLm51dGFuaXgKLnNwZWMudmVyc2lvbnNCdW5kbGVzW10ucGFja2FnZUNvbnRyb2xsZXIKLnNwZWMudmVyc2lvbnNCdW5kbGVzW10uc25vdwouc3BlYy52ZXJzaW9uc0J1bmRsZXNbXS50aW5rZXJiZWxsCi5zcGVjLnZlcnNpb25zQnVuZGxlc1tdLnVwZ3JhZGVyCi5zcGVjLnZlcnNpb25zQnVuZGxlc1tdLnZTcGhlcmU="
 )
 
 // AlwaysExcluded are fields we always exclude from signature generation.
@@ -66,7 +82,7 @@ func GetBundleSignature(ctx context.Context, bundle *anywherev1alpha1.Bundles, k
 	}
 
 	// Create KMS Client for bundle manifest signing
-	kmsClient, err := clients.CreateKMSClient(ctx)
+	kmsClient, err := CreateKMSClient(ctx)
 	if err != nil {
 		return "", fmt.Errorf("creating kms client: %v", err)
 	}
@@ -122,7 +138,7 @@ func getDigest(bundle *anywherev1alpha1.Bundles) ([32]byte, []byte, error) {
 // representation of the Bundles object using gojq.
 func filterExcludes(jsonBytes []byte) ([]byte, error) {
 	// Decode the base64-encoded excludes
-	exclBytes, err := base64.StdEncoding.DecodeString(constants.Excludes)
+	exclBytes, err := base64.StdEncoding.DecodeString(Excludes)
 	if err != nil {
 		return nil, fmt.Errorf("decoding Excludes: %v", err)
 	}
@@ -168,4 +184,15 @@ func filterExcludes(jsonBytes []byte) ([]byte, error) {
 		return nil, fmt.Errorf("marshalling final result to JSON: %v", err)
 	}
 	return filtered, nil
+}
+
+// CreateKMSClient creates KMS client for bundle manifest signing.
+func CreateKMSClient(ctx context.Context) (*kms.Client, error) {
+	conf, err := config.LoadDefaultConfig(ctx, config.WithRegion(DefaultRegion))
+	if err != nil {
+		return nil, fmt.Errorf("loading AWS config in region %q: %v", DefaultRegion, err)
+	}
+	client := kms.NewFromConfig(conf)
+
+	return client, nil
 }
